+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * **Issue #240 - JMeter version 3.3 support added.**
 
+* **Code cleanups and better logging.**
+
 ##Version 2.2.0 Release Notes
 
 * **Issue #222 - JMeter version 3.2.0 support added.**